Reconnect when Shoutcast queue becomes too large
In case of a connectivity problem with the Shoutcast server, it may take
a considerable amount of time before the client notices that and
reconnects. To speed things up, after each shout_send() we now check the
value of shout_queuelen() and reconnect if the queue is too large.

The max queue length is currently hardcoded to 64K on Win32 and 32K
otherwise. This translates to about 20 seconds.