PulseAudio: synchronize left and right channels of stereo streams
PulseAudio has special provisions to set two streams as a synchronized
pair. We utilize this mechanism to keep left and right channels of
stereo-mixed streams in sync. It is still possible for the channels
to desync a bit, however with both streams being set up as as sync
pair it happens considerably less often. The best way would be to
downmix both channels into an interleaved sample block, however
right now we don't do this for the sake of simplicity and to keep
CPU usage low.