Properly handle sampling rates being non-integer multiples of WAVE_RATE
First, we have to advance the FFT window by a whole number of samples,
(I/Q pairs). This was broken, which caused garbled audio eg. on Airspy
Mini using sampling rate of 3 Msps.

Second, the FFT window may only slide by an integer number of input
samples, while in fact required input-to-audio decimation factor is
a rational number. A non-zero rounding error introduces additional
phase rotation which we have to compensate in order to shift the
channel of interest to the center of the spectrum of the output
I/Q stream. This is important for correct NFM demodulation and
producing proper data for rawfile outputs.