SoapySDR: configurable gain components
Individual gain components of the hardware can now be configured.
This is done by setting the gain option to a string consisting of
a series of gain_component_name=value pairs separated with commas
(it is parsed with SoapySDRKwargs_fromString()). "gain" option
can still be given as an integer or floating point value - in this
case, gain is set with SoapySDRDevice_setGain() routine which
should distribute the given value to gain components supported
by the hardware.