Added -F cmdline option - run in foreground, no waterfalls
This shall be used to run the program as a systemd service
of type "simple". Some libraries (notably librtlsdr and
SoapySDR) print their diagnostic messages to stderr, which
is lost when running as a daemon. When running as a simple
service, these messages are captured by systemd and written
to the system journal.