"syslog" config option replaced with -e cmdline option
The rationale behind this is that we usually want syslog logging
to be enabled when running in foreground with TUI enabled (in order
not to clutter the waterfalls with messages). But when running
as a systemd service, messages shall be printed to stderr and
captured into system journal by systemd. Frequent flipping from
foreground to background mode and back is common during configuration
tuning and troubleshooting. In such scenarios it's more convenient
to change log destination with a command line switch rather than
config option.