Upscale I/Q samples before writing them to a file
2d412580 changed the scaling factor of the input samples and now
we are operating internally on a range between -1.0 to 1.0[1]. When
using rawfile outputs, samples must be upscaled back.

[1] In fact, samples after FFT and decimation may exceed this value
due to the fact that the channel of interest may not necessarily
be exactly centered on the configured frequency. If this is the case,
I/Q samples written into rawfile outputs are superimposed onto a
complex exponential and this stretches the range of their values.
In order not to overflow the CS16 range, samples are multiplied by
1024.0 rather than 32768, just to be on a safe side.