Dated subdirectories.
When output files collected with split_on_transmission are being
collected for a long time the output directories become inconvenient
to navigate. They can have a huge amount of individual .mp3 files.
This patch adds a configuration option that will put files in
dated subdirectories.

For example, the file:
  output_dir/prefix_20230831_204504.mp3

with the dated_subdirectories option will be saved like so:
  output_dir/2023/08/31/prefix_20230831_204504.mp3

Proposed wiki addition:
* `dated_subdirectories` (boolean, optional) - when enabled will create daily
  subdirectories under the specified directory. The directories follow the
  pattern directory/YYYY/MM/DD. The default is false.